+++ src/main/java/dev/dediamondpro/resourcify/mixins/PackScreenAccessor.java
@@ -31,5 +31,5 @@ public interface PackScreenAccessor {
     Path getDirectory();
 
     @Accessor("model")
-    PackSelectionModel getOrganizer();
+    PackSelectionModel getModel();
 }