+++ src/main/kotlin/dev/dediamondpro/resourcify/util/NetworkUtil.kt
@@ -18,6 +18,7 @@
 package dev.dediamondpro.resourcify.util
 
 import dev.dediamondpro.resourcify.ModInfo
+import dev.dediamondpro.resourcify.platform.Platform
 import gg.essential.universal.UMinecraft
 import java.awt.image.BufferedImage
 import java.io.InputStream
@@ -39,7 +40,10 @@ object NetworkUtil {
 
     fun setupConnection(url: URL): HttpsURLConnection {
         val con = url.openConnection() as HttpsURLConnection
-        con.setRequestProperty("User-Agent", "${ModInfo.NAME}/${ModInfo.VERSION}")
+        con.setRequestProperty(
+            "User-Agent",
+            "${ModInfo.NAME}/${ModInfo.VERSION} (${Platform.getMcVersion()}-${Platform.getLoader()})"
+        )
         con.setRequestProperty("Accept-Encoding", "gzip, deflate")
         con.connectTimeout = 5000
         con.readTimeout = 5000